+++ frontend/src/app/services/media.service.ts
@@ -13,5 +13,18 @@ export class MediaService {
     getAllmedia(){
           return this.http.get<MediaDTO[]>('/api/getAllMedia');
     }
+
+    openMedia(sorszam: number){
+          return this.http.get<MediaDTO>('/api/openMedia/' + sorszam);
+    }
+
+    modifyMedia(sorszam: number, cim?: string, beszerzes_datuma?: string, tipus?: string) {
+      const body: any = {};
+      if (cim !== undefined) body.cim = cim;
+      if (beszerzes_datuma !== undefined) body.beszerzes_datuma = beszerzes_datuma;
+      if (tipus !== undefined) body.tipus = tipus;
+  
+      return this.http.put('/api/modifyMedia/' + sorszam, body);
+    }
     
 }